+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- **[Fix]** Add support for relative source map URLs.
+
 # 0.3.1 (2018-30-06)
 
 - **[Feature]** Use `node-inspector-server` instead of `demurgos-spawn-wrap` to inspect subprocesses.